+++ src/main/java/com/dabomstew/pkrandom/romhandlers/Gen3RomHandler.java
@@ -3404,9 +3404,12 @@ public class Gen3RomHandler extends AbstractGBRomHandler {
         List<TypeRelationship> typeEffectivenessTable = readTypeEffectivenessTable();
         Type.STRONG_AGAINST.clear();
         Type.RESISTANT_TO.clear();
+        Type.IMMUNE_TO.clear();
         for (TypeRelationship rel : typeEffectivenessTable) {
             switch (rel.effectiveness) {
                 case ZERO:
+                    Type.updateImmuneTo(rel.attacker, rel.defender);
+                    break;
                 case HALF:
                     Type.updateResistantTo(rel.attacker, rel.defender);
                     break;
@@ -3444,9 +3447,8 @@ public class Gen3RomHandler extends AbstractGBRomHandler {
         int currentOffset = getRomEntry().getValue("TypeEffectivenessOffset");
         int attackingType = readByte(currentOffset);
         // 0xFE marks the end of the table *not* affected by Foresight, while 0xFF marks
-        // the actual end of the table. Since we don't care about Ghost immunities at all,
-        // just stop once we reach the Foresight section.
-        while (attackingType != (byte) 0xFE) {
+        // the actual end of the table.
+        while (attackingType != (byte) 0xFF) {
             Type attacking = getTypeTableValue(currentOffset);
             Type defending = getTypeTableValue(currentOffset + 1);
             int effectivenessInternal = readByte(currentOffset + 2);
@@ -3472,6 +3474,10 @@ public class Gen3RomHandler extends AbstractGBRomHandler {
             }
             currentOffset += 3;
             attackingType = readByte(currentOffset);
+            while (attackingType == (byte) 0xFE) {
+                currentOffset += 3;
+                attackingType = readByte(currentOffset);
+            }
         }
         return typeEffectivenessTable;
     }