+++ src/main/java/com/dabomstew/pkrandom/romhandlers/Gen4RomHandler.java
@@ -3048,9 +3048,12 @@ public class Gen4RomHandler extends AbstractDSRomHandler {
                         readTypeEffectivenessTable(battleOverlay, typeEffectivenessTableOffset);
                 Type.STRONG_AGAINST.clear();
                 Type.RESISTANT_TO.clear();
+                Type.IMMUNE_TO.clear();
                 for (TypeRelationship rel : typeEffectivenessTable) {
                     switch (rel.effectiveness) {
                         case ZERO:
+                            Type.updateImmuneTo(rel.attacker, rel.defender);
+                            break;
                         case HALF:
                             Type.updateResistantTo(rel.attacker, rel.defender);
                             break;
@@ -3109,9 +3112,8 @@ public class Gen4RomHandler extends AbstractDSRomHandler {
         int currentOffset = typeEffectivenessTableOffset;
         int attackingType = battleOverlay[currentOffset];
         // 0xFE marks the end of the table *not* affected by Foresight, while 0xFF marks
-        // the actual end of the table. Since we don't care about Ghost immunities at all,
-        // just stop once we reach the Foresight section.
-        while (attackingType != (byte) 0xFE) {
+        // the actual end of the table.
+        while (attackingType != (byte) 0xFF) {
             int defendingType = battleOverlay[currentOffset + 1];
             int effectivenessInternal = battleOverlay[currentOffset + 2];
             Type attacking = Gen4Constants.typeTable[attackingType];
@@ -3138,6 +3140,10 @@ public class Gen4RomHandler extends AbstractDSRomHandler {
             }
             currentOffset += 3;
             attackingType = battleOverlay[currentOffset];
+            while (attackingType == (byte) 0xFE) {
+                currentOffset += 3;
+                attackingType = battleOverlay[currentOffset];
+            }
         }
         return typeEffectivenessTable;
     }